clean up code, add 2 unit tests

+++ journal.py
@@ -10,6 +10,7 @@ class Entry:
         self.content = content
     
     def isValidEntry(self):
+        self.title, self.content = self.title.strip(), self.content.strip()
         return len(self.title) > 0 and len(self.content) > 0
 
     def getContent(self):
@@ -25,8 +26,11 @@ class Journal:
         self.myIndent = 4
     
     def addEntry(self, title, content):
+
         newEntry = Entry(title, content)
-        
+        if not newEntry.isValidEntry(): 
+            print("\nYou must enter both a title and content of length > 0\n")
+            return
         if not os.path.isfile(self.fileName):
             firstEntry = [ newEntry.__dict__ ]
             with open(self.fileName, mode='w') as f:
@@ -36,10 +40,22 @@ class Journal:
             with open(self.fileName) as f:
                 currData = json.load(f)
 
+            if self.entryExists(newEntry, currData):
+                print("\nYou already have a journal entry with the same title\n")
+                return
+
             currData.append(newEntry.__dict__)
 
             with open(self.fileName, mode='w') as f:
                 f.write(json.dumps(currData, indent=self.myIndent))
+
+    # O(n), checks if there is a journal entry with the same name as the new one a user wants to add
+    def entryExists(self, entry, currData):
+        for i in range(len(currData)):
+                currEntry = currData[i]
+                if currEntry['title'] == entry.title:
+                    return True
+        return False
 
     def printJournalTitles(self):
         if not os.path.isfile(self.fileName):
@@ -59,31 +75,33 @@ class Journal:
     # prints the contents of the json file, used for testing
     def printJournal(self):
         if not os.path.isfile(self.fileName):
-            print("No entries in journal, nothing to print")
+            print("\nNo entries in journal, nothing to print\n")
         else:
             with open(self.fileName) as f:
                 data = f.read()
                 print(data)    
 
-def main():
+def main(myJournal):
     parser = argparse.ArgumentParser(description="Journal Entry Tracker")
     group = parser.add_mutually_exclusive_group()
 
     group.add_argument("-c", "--create", help="the contents")
-    parser.add_argument("--title", help="the title")
+    parser.add_argument("-t", "--title", help="the title")
     group.add_argument("-l", "--list", help="List entries", action="store_true")
 
     args = parser.parse_args()
 
     if args.list:
         myJournal.printJournalTitles()
-    elif args.create:
+    elif args.create and args.title:
         myJournal.addEntry(args.title, args.create)
     else:
-        print("You must povide an action: create entry or list entries")
-
+        print("\nYou must provide an action: create entry with both a title and content or list entries. Use --help for help.\n")
+        # raise RuntimeError('You must provide an action: create entry with both a title and content or list entries. Use --help for help.')
+        
 if __name__ == '__main__':
     myJournal = Journal("data.json")
-    main()
+    main(myJournal)
+
